+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Beginner grows its eco bonus slower
 - Casual through Platinum start gaining AI tech slightly later
 - Beginner and Casual spawning additional Commanders if selected after a higher difficulty
+- Cluster is disabled if TITANS is not enabled
 
 ## v5.35.0 - 2022-06-15